# api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: SundayBest API
  version: 1.0.0
servers:
  - url: http://localhost:4100
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 200 }
    ClientTimezone:
      name: X-Client-Timezone
      in: header
      required: false
      schema: { type: string, example: America/New_York }
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
paths:
  /health/live:
    get:
      operationId: healthLive
      responses:
        '200': { description: Alive }
  /health/ready:
    get:
      operationId: healthReady
      responses:
        '200': { description: PostgreSQL reachable }
        '503': { description: Not ready }
  /v1/attest/challenge:
    post:
      operationId: createAttestChallenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { keyId: { type: string } }
      responses:
        '200': { description: One-time challenge }
  /v1/attest/verify:
    post:
      operationId: verifyAttestation
      responses:
        '200': { description: New anonymous session }
  /v1/session/bootstrap:
    post:
      operationId: bootstrapSession
      responses:
        '200': { description: Session restored from an attested key }
  /v1/session/refresh:
    post:
      operationId: refreshSession
      responses:
        '200': { description: Rotated session credentials }
  /v1/session/revoke:
    post:
      security: [{ bearerAuth: [] }]
      operationId: revokeSession
      responses:
        '200': { description: Session revoked }
  /v1/me:
    get:
      security: [{ bearerAuth: [] }]
      operationId: getMe
      responses: { '200': { description: Current user } }
    patch:
      security: [{ bearerAuth: [] }]
      operationId: updateMe
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '200': { description: Updated user } }
    delete:
      security: [{ bearerAuth: [] }]
      operationId: deleteMe
      description: Requires request-bound App Attest assertion headers.
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '200': { description: User data deleted } }
  /v1/me/onboarding/complete:
    post:
      security: [{ bearerAuth: [] }]
      operationId: completeOnboarding
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '200': { description: Onboarding completed } }
  /v1/me/settings:
    get:
      security: [{ bearerAuth: [] }]
      operationId: getSettings
      responses: { '200': { description: Settings } }
    patch:
      security: [{ bearerAuth: [] }]
      operationId: updateSettings
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '200': { description: Settings updated } }
  /v1/me/reminders:
    get:
      security: [{ bearerAuth: [] }]
      operationId: listReminders
      responses: { '200': { description: Reminder preferences } }
  /v1/me/reminders/{kind}:
    put:
      security: [{ bearerAuth: [] }]
      operationId: updateReminder
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - name: kind
          in: path
          required: true
          schema: { type: string, enum: [dailyStudy, quickCheck] }
      responses: { '200': { description: Reminder updated } }
  /v1/sermons/resolve:
    post:
      security: [{ bearerAuth: [] }]
      operationId: resolveSermon
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '200': { description: Resolved YouTube sermon metadata } }
  /v1/plans:
    get:
      security: [{ bearerAuth: [] }]
      operationId: listPlans
      responses: { '200': { description: User-visible plans } }
    post:
      security: [{ bearerAuth: [] }]
      operationId: createPlan
      description: Requires request-bound App Attest assertion headers; queues asynchronous generation.
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '200': { description: Plan generation accepted } }
  /v1/plans/{planId}:
    get:
      security: [{ bearerAuth: [] }]
      operationId: getPlan
      responses: { '200': { description: Plan detail } }
  /v1/plans/{planId}/start:
    post:
      security: [{ bearerAuth: [] }]
      operationId: startPlan
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/ClientTimezone' }
      responses: { '200': { description: Started plan with scheduled days } }
  /v1/plans/{planId}/archive:
    post:
      security: [{ bearerAuth: [] }]
      operationId: archivePlan
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '200': { description: Archived plan } }
  /v1/plans/{planId}/saved:
    put:
      security: [{ bearerAuth: [] }]
      operationId: savePlan
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '200': { description: Saved } }
    delete:
      security: [{ bearerAuth: [] }]
      operationId: unsavePlan
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '200': { description: Removed from saved } }
  /v1/plan-generations/{generationId}:
    get:
      security: [{ bearerAuth: [] }]
      operationId: getGeneration
      responses: { '200': { description: Generation status } }
  /v1/plan-generations/{generationId}/retry:
    post:
      security: [{ bearerAuth: [] }]
      operationId: retryGeneration
      description: Requires request-bound App Attest assertion headers.
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '200': { description: Failed generation queued again } }
  /v1/plans/{planId}/days/{dayNumber}:
    get:
      security: [{ bearerAuth: [] }]
      operationId: getStudyDay
      parameters: [{ $ref: '#/components/parameters/ClientTimezone' }]
      responses: { '200': { description: Study day in user's current Bible translation } }
  /v1/plans/{planId}/days/{dayNumber}/steps/{step}:
    put:
      security: [{ bearerAuth: [] }]
      operationId: completeStudyStep
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/ClientTimezone' }
      responses: { '200': { description: Step completed } }
  /v1/plans/{planId}/days/{dayNumber}/complete:
    post:
      security: [{ bearerAuth: [] }]
      operationId: completeStudyDay
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/ClientTimezone' }
      responses: { '200': { description: Day completed } }
  /v1/quizzes/{quizId}/attempts:
    post:
      security: [{ bearerAuth: [] }]
      operationId: startQuizAttempt
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/ClientTimezone' }
      responses: { '200': { description: Quiz + attempt without answer keys } }
  /v1/quiz-attempts/{attemptId}:
    get:
      security: [{ bearerAuth: [] }]
      operationId: getQuizAttempt
      responses: { '200': { description: Attempt state } }
  /v1/quiz-attempts/{attemptId}/answers:
    post:
      security: [{ bearerAuth: [] }]
      operationId: submitQuizAnswer
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '200': { description: Immediate answer feedback } }
  /v1/quiz-attempts/{attemptId}/complete:
    post:
      security: [{ bearerAuth: [] }]
      operationId: completeQuizAttempt
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '200': { description: Completed attempt + score } }